Adding timeout on context for all commands
