+++ ee/acl/acl.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
+	"time"
 
 	"github.com/dgraph-io/dgo"
 	"github.com/dgraph-io/dgo/protos/api"
@@ -94,7 +95,8 @@ func userAdd(conf *viper.Viper, userid string, password string) error {
 		}
 	}
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -132,7 +134,8 @@ func groupAdd(conf *viper.Viper, groupId string) error {
 	}
 	defer cancel()
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -204,7 +207,8 @@ func userOrGroupDel(conf *viper.Viper, userOrGroupId string,
 	}
 	defer cancel()
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -253,16 +257,18 @@ func mod(conf *viper.Viper) error {
 			return err
 		}
 
-		if (conf.GetBool("new_password") && conf.GetString("group_list") != defaultGroupList) ||
-			(!conf.GetBool("new_password") && conf.GetString("group_list") == defaultGroupList) {
+		newPassword := conf.GetBool("new_password")
+		groupList := conf.GetString("group_list")
+		if (newPassword && groupList != defaultGroupList) ||
+			(!newPassword && groupList == defaultGroupList) {
 			return fmt.Errorf("one of --new_password or --group_list must be provided, but not both")
 		}
 
-		if conf.GetBool("new_password") {
+		if newPassword {
 			return changePassword(conf, userId)
 		}
 
-		return userMod(conf, userId, conf.GetString("group_list"))
+		return userMod(conf, userId, groupList)
 	}
 
 	// when modifying the group, some user options are forbidden
@@ -287,7 +293,8 @@ func changePassword(conf *viper.Viper, userId string) error {
 		return err
 	}
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -329,7 +336,8 @@ func userMod(conf *viper.Viper, userId string, groups string) error {
 	}
 	defer cancel()
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -423,7 +431,8 @@ func chMod(conf *viper.Viper) error {
 	}
 	defer cancel()
 
-	ctx := context.Background()
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
 	txn := dc.NewTxn()
 	defer func() {
 		if err := txn.Discard(ctx); err != nil {
@@ -588,4 +597,82 @@ func updateAcl(acls []Acl, newAcl Acl) ([]Acl, bool) {
 
 	// we do not find any existing aclEntry matching the newAcl predicate
 	return append(acls, newAcl), true
+}
+
+func queryAndPrintUser(ctx context.Context, txn *dgo.Txn, userId string) error {
+	user, err := queryUser(ctx, txn, userId)
+	if err != nil {
+		return err
+	}
+	if user == nil {
+		return fmt.Errorf("The user %q does not exist.\n", userId)
+	}
+
+	fmt.Printf("User  : %s\n", userId)
+	fmt.Printf("UID   : %s\n", user.Uid)
+	for _, group := range user.Groups {
+		fmt.Printf("Group : %-5s\n", group.GroupID)
+	}
+	return nil
+}
+
+func queryAndPrintGroup(ctx context.Context, txn *dgo.Txn, groupId string) error {
+	group, err := queryGroup(ctx, txn, groupId, "dgraph.xid", "~dgraph.user.group{dgraph.xid}",
+		"dgraph.group.acl")
+	if err != nil {
+		return err
+	}
+	if group == nil {
+		return fmt.Errorf("The group %q does not exist.\n", groupId)
+	}
+	fmt.Printf("Group: %s\n", groupId)
+	fmt.Printf("UID  : %s\n", group.Uid)
+	fmt.Printf("ID   : %s\n", group.GroupID)
+
+	var userNames []string
+	for _, user := range group.Users {
+		userNames = append(userNames, user.UserID)
+	}
+	fmt.Printf("Users: %s\n", strings.Join(userNames, " "))
+
+	var acls []Acl
+	if len(group.Acls) != 0 {
+		if err := json.Unmarshal([]byte(group.Acls), &acls); err != nil {
+			return fmt.Errorf("unable to unmarshal the acls associated with the group %v: %v",
+				groupId, err)
+		}
+
+		for _, acl := range acls {
+			fmt.Printf("ACL  : %v\n", acl)
+		}
+	}
+	return nil
+}
+
+func info(conf *viper.Viper) error {
+	userId, groupId, err := getUserAndGroup(conf)
+	if err != nil {
+		return err
+	}
+
+	dc, cancel, err := getClientWithAdminCtx(conf)
+	defer cancel()
+	if err != nil {
+		return fmt.Errorf("unable to get admin context: %v\n", err)
+	}
+
+	ctx, ctxCancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer ctxCancel()
+	txn := dc.NewTxn()
+	defer func() {
+		if err := txn.Discard(ctx); err != nil {
+			fmt.Printf("Unable to discard transaction: %v\n", err)
+		}
+	}()
+
+	if len(userId) != 0 {
+		return queryAndPrintUser(ctx, txn, userId)
+	}
+
+	return queryAndPrintGroup(ctx, txn, groupId)
 }
